psychologist_app/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
werkzeug==3.0.1
eventlet==0.33.3
openai>=1.12.0
python-dotenv==1.0.0
selenium==4.15.2
psycopg2-binary==2.9.9
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.0
pyotp==2.9.0
qrcode[pil]==7.4.2
requests==2.31.0

